+++ indexer/modules/custom/merchant_moe/domain/merchant_moe.py
@@ -30,3 +30,21 @@ class MerChantMoePool(FilterData):
     token1_address: str
     block_number: int
     block_timestamp: int
+
+
+@dataclass
+class MerchantMoeErc1155TokenSupply(FilterData):
+    token_address: str
+    token_id: int
+    total_supply: int
+    block_number: int
+    block_timestamp: int
+
+
+@dataclass
+class MerchantMoeErc1155TokenCurrentSupply(FilterData):
+    token_address: str
+    token_id: int
+    total_supply: int
+    block_number: int
+    block_timestamp: int
