--- indexer/modules/custom/merchant_moe/domain/merchant_moe_1155_total_supply.py
@@ -1,21 +1,0 @@
-from dataclasses import dataclass
-
-from indexer.domain import Domain, FilterData
-
-
-@dataclass
-class MerchantMoeErc1155TokenSupply(FilterData):
-    token_address: str
-    token_id: int
-    total_supply: int
-    block_number: int
-    block_timestamp: int
-
-
-@dataclass
-class MerchantMoeErc1155TokenCurrentSupply(FilterData):
-    token_address: str
-    token_id: int
-    total_supply: int
-    block_number: int
-    block_timestamp: int
